+++ plugin/account-import/src/main/java/org/zstack/identity/imports/entity/ThirdPartyAccountSourceVO.java
@@ -26,7 +26,7 @@ public class ThirdPartyAccountSourceVO extends ResourceVO {
     private String description;
 
     /**
-     * ex: "LDAP"
+     * ex: "ldap"
      */
     @Column
     private String type;
